Changed some size_t's
date	99.04.07.16.53.00;	author grsmith1;	state Exp;

+++ source/components/utilities/utdebug.c
@@ -566,7 +566,7 @@ cleanup:
 
 
 UINT8
-checksum (void *pvBuf,size_t stBufSize)
+checksum (void *pvBuf,UINT32 stBufSize)
 {
     return 0;
 }
@@ -588,6 +588,7 @@ IncIndent (void)
 {
     return 0;
 }
+
